pyproject.toml
[tool.poetry]
name = "stock price tracker"
version = "0.1.0"
description = ""
authors = ["Chuan <kirk336neo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.1.1"
numpy = "^1.26.1"
pygsheets = "^2.0.6"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
